fix: 1등이 여러번 당첨될 경우 int의 범위를 넘어가기에 long으로 변경

+++ src/main/java/lotto/controller/LottoController.java
@@ -79,7 +79,7 @@ public class LottoController {
 
     private void printStatistics() {
         final Map<Winning, Integer> winningCounts = lottoService.getWinningCounts();
-        int totalPrize = lottoService.calculateTotalPrize(winningCounts);
+        long totalPrize = lottoService.calculateTotalPrize(winningCounts);
         outputView.printWinningStatistics();
         outputView.printStatistics(winningCounts, totalPrize, user.getPurchaseAmount().amount());
     }
